Rearranged functions for better understanding of codes
